[Fix] XM Compatibility: A somewhat more compatible Rxy effect. Still needs more testing, as it's approximately the same algorithm as in Milky (which I know is not 100% correct). [Fix] Comments tab: The lower part of the tab was receiving update messages that were not even meaningful to this tab (f.e. speed changes), so it was updating quite often modules that have alternating speed... Only letting in important update messages now. [Fix] Wave Convert: The wave convert dialog had two default buttons, one of them (the "default default button") didn't make sense.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@440 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Sndfile.h
@@ -220,7 +220,7 @@ typedef struct __declspec(align(32)) _MODCHANNEL
 	BYTE nOldCmdEx, nOldVolParam, nOldTempo;
 	BYTE nOldOffset, nOldHiOffset;
 	BYTE nCutOff, nResonance;
-	BYTE nRetrigCount, nRetrigParam;
+	int nRetrigCount, nRetrigParam;
 	BYTE nTremorCount, nTremorParam;
 	BYTE nPatternLoop, nPatternLoopCount;
 	BYTE nRowNote, nRowInstr;
@@ -846,7 +846,7 @@ private:
 	void FineVolumeDown(MODCHANNEL *pChn, UINT param);
 	void Tremolo(MODCHANNEL *pChn, UINT param);
 	void Panbrello(MODCHANNEL *pChn, UINT param);
-	void RetrigNote(UINT nChn, UINT param, UINT offset=0);  //rewbs.volOffset: added last param
+	void RetrigNote(UINT nChn, int param, UINT offset=0);  //rewbs.volOffset: added last param
 	void SampleOffset(UINT nChn, UINT param, bool bPorta);	//rewbs.volOffset: moved offset code to own method
 	void NoteCut(UINT nChn, UINT nTick);
 	int PatternLoop(MODCHANNEL *, UINT param);
